+++ src/node-commands.ts
@@ -69,6 +69,7 @@ export function preflightNPMCheck(project: Project): string {
 export function npmInstallAll(): string {
   switch (ionicState.repoType) {
     case MonoRepoType.pnpm:
+    case MonoRepoType.lerna:
     case MonoRepoType.folder:
       return InternalCommand.cwd + pm(PMOperation.installAll);
     default:
@@ -134,7 +135,7 @@ export function npmUninstall(name: string): string {
       return `${pm(PMOperation.uninstall, name)} --workspace=${ionicState.workspace}`;
     case MonoRepoType.folder:
     case MonoRepoType.yarn:
-      case MonoRepoType.lerna:
+    case MonoRepoType.lerna:
     case MonoRepoType.pnpm:
       return `${InternalCommand.cwd}${pm(PMOperation.uninstall, name)}`;
     default:
@@ -148,7 +149,7 @@ export function npmRun(name: string): string {
       return `${pm(PMOperation.run, name)} --workspace=${ionicState.workspace}`;
     case MonoRepoType.folder:
     case MonoRepoType.yarn:
-      case MonoRepoType.lerna:
+    case MonoRepoType.lerna:
     case MonoRepoType.pnpm:
       return `${InternalCommand.cwd}${pm(PMOperation.run, name)}`;
     default: