+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Report deprecated packages and plugins
 - Report warning on remote dependencies
 - Recommend removal of rxjs-compat (as you should migrate v5 to v6+)
+- Report Cordova plugins in a Cordova project that cannot be inspected for Capacitor compatibility
 
 ### Version 1.5.13